requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
pyjwt
bcrypt